% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Azov.R
\docType{data}
\name{Azov}
\alias{Azov}
\title{Helminth communities of so-iuy mullets from the Sea of Azov}
\format{A data frame with 378 rows and 26 variables:
\describe{
  \item{sample}{survey identifier}
  \item{Ssp1}{abundance (no. individuals) of Solostamenides sp. 1}
  ...
}}
\source{
\url{https://doi.org/10.7910/DVN/IWIKOL}
}
\usage{
data(grav)
}
\description{
A dataset containing the abundance of 25 helminth species from 378 so-iuy
mullets collected in the Sea of Azov and the Black Sea. Fish are grouped
in seven surveys
}
\details{
#' @docType data
}
\references{
Llopis‐Belenguer, C., Blasco‐Costa, I., Balbuena, J.A.,
Sarabeev, V. and Stouffer, D.B. (2020), Native and invasive hosts play
different roles in host–parasite networks. Ecography, 43: 559-568.
doi:10.1111/ecog.04963
}
\keyword{datasets}
