% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{AOplot}
\alias{AOplot}
\title{Abundance Occupancy Plot}
\usage{
AOplot(
  fq,
  col.rc = c("red", "blue"),
  opacity = 0.1,
  log.x = FALSE,
  log.y = FALSE,
  xLab = "Fraction of sites occupied",
  yLab = "Mean abundance",
  ...
)
}
\arguments{
\item{fq}{a fuzzyq object returned by function fuzzyq}

\item{col.rc}{a vector specifying two colors to be used for plot. Accepts any valid color specification in R.}

\item{opacity}{number between 0 and 1 specificying the opacity of convex hulls groupping common and rare species.}

\item{log.x}{logical flag indicating whether the x axis show be in log10 scale.}

\item{log.y}{logical flag indicating whether the y axis show be in log10 scale.}

\item{xLab}{a title for the x axis.}

\item{yLab}{a title for the y axis}

\item{...}{arguments to be passed to graphical parameters in R.}
}
\value{
a scatter plot of occupancy vs. abundance of species. Convex hulls identify common and rare species.
}
\description{
Plots the abundance-occupancy relationship of species in a community categorized as common or rare by fuzzyq.
}
\examples{
data(antsA)
# Data set of 46 ant species colelcted in 100 sites.
FQAnts <- fuzzyq(antsA, sorting = TRUE)
AOplot(FQAnts) # Plor with default values
# Alternative with colors specified in Hex format, logarithmic axes and point format
AOplot(FQAnts, col.rc = c("#013bad","#bd5f69"),
       log.x = TRUE, log.y = TRUE, pch =16)
}
