% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyq.R
\name{fuzzyq}
\alias{fuzzyq}
\title{Fuzzy Quantification of Common and Rare Species in Ecological Communities}
\usage{
fuzzyq(
  M,
  nclus = 2,
  diss = "gower",
  rm.absent = FALSE,
  sorting = TRUE,
  keep.Diss = FALSE,
  daisy.args,
  ...
)
}
\arguments{
\item{M}{a Matrix or dataframe of species abundaces (columns). Each row represents a sites or sampling unit.}

\item{nclus}{number of clusters into which species should be allocated. Default is 2 (common and rare).}

\item{diss}{string specifiying the dissimilarity coefficient to be used. Default is "gower". The other options are "euclidean" and "manhattan".}

\item{rm.absent}{a logical flag that indicates how to treat species absent in the community. If TRUE absences are considered as structural; absent species are removed from calculations. If FALSE (the default) absences are considered as random; absent species are included in the calculations.}

\item{sorting}{logical flag: If TRUE (the default) species are sorted in the output by ascending silhouette width within each cluster, else species are arranged in the same order as in the input.}

\item{keep.Diss}{logical flag specifying whether the species dissimilarity matrix shoudl be returned. The default is FALSE.}

\item{daisy.args}{arguments to be passed to function daisy in package cluster.}

\item{...}{arguments to be passed to function fanny in package cluster.}
}
\value{
a list of class fuzzyq with the abundance occupancy of each species, clustering metrics for each species and the whole community, and the species dissimilary matrix (if diss = TRUE).
}
\description{
Perform fuzzy clustering of each species based on its abundance and occupancy.
}
\examples{
data(antsA)
# Data set of 46 ant species colelcted in 100 sites.
FQAnts <- fuzzyq(antsA, sorting = TRUE)
}
