% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyq.R
\name{fuzzyq}
\alias{fuzzyq}
\title{Fuzzy Quantification of Common and Rare Species in Ecological Communities}
\usage{
fuzzyq(
  M,
  nclus = 2,
  diss = "gower",
  rm.absent = FALSE,
  sorting = TRUE,
  keep.Diss = FALSE,
  daisy.args,
  ...
)
}
\arguments{
\item{M}{A matrix or dataframe of species abundaces (columns). Each row represents a site.}

\item{nclus}{Integer. Indicate the number of clusters into which species should be allocated.
Default is 2 (common and rare).}

\item{diss}{String. Specifiy the dissimilarity coefficient to be used. Default is "gower".
See \code{\link[cluster]{daisy}} for other choices.}

\item{rm.absent}{Logical. Whether or not absent species are to be removed from the calculations.}

\item{sorting}{Logical. If \code{TRUE} (the default) species are sorted in the output by ascending
silhouette widths within each cluster, else species are arranged in the same order as in the
input matrix or dataframe.}

\item{keep.Diss}{Logical. Whether or not the species dissimilarity matrix shoudl be returned. The
default is \code{FALSE}.}

\item{daisy.args}{Arguments to be passed to function \code{daisy} in package \code{cluster}.}

\item{...}{Arguments to be passed to function \code{fanny} in package \code{cluster}.}
}
\value{
A list of class \code{fuzzyq} containing the following:
\itemize{
  \item \code{A_O}{Abundance-occupancy information for each species.}
  \item \code{Diss}{Object of class dist with pairwise dissimilarities among species based on A_O.
  (only if \code{keep.Diss = TRUE)}.}
  \item \code{spp}{Clustering metrics per species: Cluster membership (where 0 and 1 denote allocation
   to the rare and common category, respectively), Silhouette Widths and Commonness Indices).}
  \item \code{global}{Community level clustering metrics: Average silhouette widths per cluster and globally,
   Mean commonness indices per cluster and Normalized Dunn's coefficient.}
}
See Also
\code{fanny} and \code{daisy} within \code{cluster}
}
\description{
Perform fuzzy clustering of each species based on its abundance and occupancy.
}
\examples{
data(antsA)
FQAnts <- fuzzyq(antsA, sorting = TRUE)
}
