% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{fuzzyqBoot}
\alias{fuzzyqBoot}
\title{Apply Fuzzy Quantification of Common and Rare Species to Bootstrap Replicates}
\usage{
fuzzyqBoot(M, N = 1000, level = "spp", rm.absent = FALSE, daisy.args, ...)
}
\arguments{
\item{M}{A matrix or dataframe of species abundaces (columns). Each row represents a site.}

\item{N}{Integer. Number of bootsrap replicates desired. Default is 1000.}

\item{level}{String. Specifiy the type of metrics to be computed for each bootstrap replicate.
Either \code{"spp"} or \code{"global"}, corresponding to species or community-level metrics,
respectively.}

\item{rm.absent}{Logical. Whether or not absent species are to be removed from the calculations.}

\item{daisy.args}{Arguments to be passed to function \code{daisy} in package \code{cluster}.}

\item{...}{Arguments to be passed to function \code{fanny} in package \code{cluster}.}

\item{sorting}{Logical. If \code{TRUE} (the default) species are sorted in the output by ascending
silhouette widths within each cluster, else species are arranged in the same order as in the
input matrix or dataframe.}
}
\value{
A list consisting of the following:
\describe{
  \item \code{bs.rep}{matrix of estimated metrics. Replicates are arranged in rows. If \code{level = "spp"},
    columns represent estimates of Commonness Indices per species. If \code{level = "global"}, columns
    represent estimates of Community level clustering metrics: Average silhouette widths per cluster
    and globally, Mean commonness indices per cluster and Normalized Dunn's coefficient.}
 \item \code{level}{Indicates whether the estimates are taken at species (\code{"spp"}) or
    community level (\code{"global"}).}
}
}
\description{
Produce N replicates of the original site by species matrix or dataframe by taking bootstrap
    samples of sites (rows) and apply \code{fuzzyq} to each replicate.
}
\examples{
data(antsA)
FQAnts <- fuzzyq(antsA, sorting = TRUE)
# Estimates of species Commonness Indices by species of 1,000 bootstrap replicates:
BS.FQAnts <- fuzzyqBoot (antsA, N = 1e3, level='spp')
# Estimates of global metrics of 1,000 boostrap replicates:
BS.global <- fuzzyqBoot (antsA, N = 1e3, level='global')
}
