% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sortClus}
\alias{sortClus}
\title{Sort Species by fuzzyq Clustering}
\usage{
sortClus(M, fq)
}
\arguments{
\item{M}{A matrix or dataframe with information of species in columns.}

\item{fq}{A list of class \code{fuzzyq} returned by \code{FuzzyQ::fuzzyq}.}
}
\value{
A matrix or dataframe with information of species in columns sorted according to
   \code{fq$spp}.
}
\description{
Sort species in a matrix or dataframe according to the clustering of a \code{fuzzyq} object. This is
    useful prior to plotting Commonness Indices derived from bootstrap replicates.
}
\examples{
data(antsA)
FQAnts <- fuzzyq(antsA, sorting = TRUE)
# Compute 95\% confidence intervals for Commonness Indices of species:
Nboot = 1e3
BS.FQAnts <- fuzzyqBoot (antsA, Nboot, level='spp')
BS.FQAnts <- fuzzyqCI(BS.FQAnts, fq=FQAnts)
# Plot Commonness Indices and their respective confidence intervals:
BS.FQAnts <- sortClus(BS.FQAnts, FQAnts)
spp <- FQAnts$spp
plot(spp[,3], cex.axis=0.8, xaxt='n', ylab="Commoness index",
    ylim=c(0, max(BS.FQAnts)), xlab="Species",col=col.RC[spp[,1]+1],
    pch=16, cex=0.8, las = 1)
ebar.int <- seq_len(nrow(spp))
arrows(ebar.int, BS.FQAnts["Lower", ], ebar.int, BS.FQAnts["Upper", ],
      length= 0, col=col.RC[spp[,1]+1])
axis(1, at=ebar.int, labels=rownames(spp), las=2, cex.axis=0.6)
}
